Support non timezone aware dates
Despite the spec [1] saying dates are in the form

    /Date(1297286970531-0800)/

The API is returning dates without timezones in the form

    /Date(1297286970531)/

Updated code to parse either format.

Dates without a timezone are made timezone aware and
assumed to be in UTC. I am not sure if this is correct,
but allows them to be compared with timezone aware dates.

Unknown date formats return None, I'm not sure if this
is correct either.

Added tests to cover all three cases.

[1] https://api.gettyimages.com/sites/default/files/GettyImagesConnectAPI.pdf